+++ src/ssw/states/stEngineFission.java
@@ -94,7 +94,7 @@ public class stEngineFission implements ifEngine, ifState {
     }
 
     public float GetCost( int MechTonnage, int Rating ) {
-        return ( 7500 * MechTonnage * Rating ) / 75;
+        return ( 7500.0f * (float) MechTonnage * (float) Rating ) / 75.0f;
     }
     
     public AvailableCode GetAvailability() {
