+++ src/ssw/states/stEngineFuelCell.java
@@ -94,7 +94,7 @@ public class stEngineFuelCell implements ifEngine, ifState {
     }
 
     public float GetCost( int MechTonnage, int Rating ) {
-        return ( 3500 * MechTonnage * Rating ) / 75;
+        return ( 3500.0f * (float) MechTonnage * (float) Rating ) / 75.0f;
     }
     
     public AvailableCode GetAvailability() {
