+++ src/ssw/states/stEngineFusion.java
@@ -94,7 +94,7 @@ public class stEngineFusion implements ifEngine, ifState {
     }
 
     public float GetCost( int MechTonnage, int Rating ) {
-        return ( 5000 * MechTonnage * Rating ) / 75;
+        return ( 5000.0f * (float) MechTonnage * (float) Rating ) / 75.0f;
     }
     
     public AvailableCode GetAvailability() {
