+++ src/ssw/states/stEngineISLF.java
@@ -90,7 +90,7 @@ public class stEngineISLF implements ifEngine, ifState {
     }
 
     public float GetCost( int MechTonnage, int Rating ) {
-        return ( 15000 * MechTonnage * Rating ) / 75;
+        return ( 15000.0f * (float) MechTonnage * (float) Rating ) / 75.0f;
     }
     
     public AvailableCode GetAvailability() {
