+++ src/ssw/states/stEngineISXXL.java
@@ -91,7 +91,7 @@ public class stEngineISXXL implements ifEngine, ifState {
     }
 
     public float GetCost( int MechTonnage, int Rating ) {
-        return ( 100000 * MechTonnage * Rating ) / 75;
+        return ( 100000.0f * (float) MechTonnage * (float) Rating ) / 75.0f;
     }
     
     public AvailableCode GetAvailability() {
